+++ e2e/cli-e2e/mocks/fixtures/code-pushup.config.js
@@ -1,3 +1,5 @@
+import { join } from 'node:path';
+import coveragePlugin from '@code-pushup/coverage-plugin';
 import eslintPlugin from '@code-pushup/eslint-plugin';
 
 export default {
@@ -20,8 +22,23 @@ export default {
         { type: 'group', plugin: 'eslint', slug: 'suggestions', weight: 1 },
       ],
     },
+    {
+      slug: 'code-coverage',
+      title: 'Code coverage',
+      refs: [
+        {
+          type: 'group',
+          plugin: 'coverage',
+          slug: 'coverage',
+          weight: 1,
+        },
+      ],
+    },
   ],
   plugins: [
     await eslintPlugin({ eslintrc: '.eslintrc.json', patterns: '**/*.ts' }),
+    await coveragePlugin({
+      reports: [join('e2e', 'cli-e2e', 'mocks', 'fixtures', 'lcov.info')],
+    }),
   ],
 };